Added Home Page with Recycler View
The home page has been added with a non-functional bottom navigation bar. I have also added a recycler view for more informative cards or tabs that users will meet on the home page with some short descriptions and images to give the user more information about the tab they are clicking.

This addition will make the changes in made earlier obsolete because they were some unnecessary imports that caused the app to crash repeatedly and without loading the home page.